[Dx12] Added support for bindless textures

+++ Arc/src/Arc/Renderer/Texture.h
@@ -11,7 +11,7 @@ namespace ArcEngine
 		
 		[[nodiscard]] virtual uint32_t GetWidth() const = 0;
 		[[nodiscard]] virtual uint32_t GetHeight() const = 0;
-		[[nodiscard]] virtual uint64_t GetRendererID() const = 0;
+		[[nodiscard]] virtual uint32_t GetRendererID() const = 0;
 		[[nodiscard]] virtual const std::string& GetPath() const = 0;
 
 		virtual void SetData(TextureData data, [[maybe_unused]] uint32_t size) = 0;
