+++ Arc/src/Platform/OpenGL/OpenGLTexture.h
@@ -19,7 +19,7 @@ namespace ArcEngine
 
 		[[nodiscard]] uint32_t GetWidth() const override { return m_Width; }
 		[[nodiscard]] uint32_t GetHeight() const override { return m_Height; }
-		[[nodiscard]] uint64_t GetRendererID() const override { return m_RendererID; }
+		[[nodiscard]] uint32_t GetRendererID() const override { return m_RendererID; }
 		[[nodiscard]] const std::string& GetPath() const override { return m_Path; }
 
 		void SetData(void* data, [[maybe_unused]] uint32_t size) override;
